src/main/python/coherent-data/requirements.txt
pandas==1.5.3
numpy==1.26.4
pillow==10.2.0
pydicom==3.0.0